% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcl6exons}
\alias{bcl6exons}
\title{Exon structure of BCL6 transcripts}
\format{GRangesList}
\source{
EnsDb.Hsapiens.v86 package
}
\usage{
bcl6exons
}
\description{
Exon structure of BCL6 transcripts
}
\examples{
bcl6exons
}
\keyword{datasets}
