% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{exonsBySJ}
\alias{exonsBySJ}
\title{Given a list of exons returns candidates for primer design}
\usage{
exonsBySJ(exonsByTx, tolerance = 0)
}
\arguments{
\item{exonsByTx}{exon GRanges by transcripts}

\item{tolerance}{the minimal tolerable difference in splice junctions 
coordinates}
}
\value{
a GRangesList
}
\description{
The transcripts of single exon are excluded!
}
