% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{exonSeqs}
\alias{exonSeqs}
\title{Return sequences for the exon pairs}
\usage{
exonSeqs(exonPairs, src)
}
\arguments{
\item{exonPairs}{a list of GRanges with the splice junction exon pairs.
For a single pair a single GRanges object may be provided.}

\item{src}{whether a BSgenome instance, e.g. BSgenome.Hsapiens.NCBI.GRCh38, 
XStringSet object or a FaFile object}
}
\value{
a list of GRanges with additional `seq` data column with
  the retrieved sequence.
}
\description{
Return sequences for the exon pairs
}
