% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{extractExonPairs}
\alias{extractExonPairs}
\title{Get exon pairs}
\usage{
extractExonPairs(exonsByTx)
}
\arguments{
\item{exonsByTx}{a list of exons returned by ensembl exonsBy tx}
}
\value{
a list of character vectors
}
\description{
Get exon pairs
}
