% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{p3Settings}
\alias{p3Settings}
\title{Get/define Primer3 settings}
\usage{
p3Settings(settings = NULL, defaultsFile = NULL)
}
\arguments{
\item{settings}{user provided list of non-default settings}

\item{defaultsFile}{(optional) a path to the file with default settings
(e.g. "primer3/primer3_v1_1_4_default_settings.txt")}
}
\value{
a named list of the settings
}
\description{
Get/define Primer3 settings
}
