% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSegments}
\alias{plotSegments}
\title{Plot exon structure}
\usage{
plotSegments(ex, colours = NULL)
}
\arguments{
\item{ex}{exon structure}

\item{colours}{colours}
}
\description{
Plot exon structure
}
