% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{primersToList}
\alias{primersToList}
\title{Split primers GRanges into a list by their direction}
\usage{
primersToList(grPrimers)
}
\arguments{
\item{grPrimers}{a GRanges with the forward and reverse primers}
}
\value{
a list with separate GRanges for every direction
}
\description{
Split primers GRanges into a list by their direction
}
