% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{primerTm}
\alias{primerTm}
\title{Set ptimer melting temperatures}
\usage{
primerTm(settings, min = NULL, optimal = NULL, max = NULL)
}
\arguments{
\item{settings}{a settings list created by `p3Settings`}

\item{min}{minimal temperature (PRIMER_MAX_TM)}

\item{optimal}{optimal temperature (PRIMER_OPT_TM)}

\item{max}{maximal temperature (PRIMER_MIN_TM)}
}
\value{
a settings list with updated values
}
\description{
Set ptimer melting temperatures
}
\examples{
p3Settings() \%>\%
  primerTm(min = 58, optimal = 60, max = 62) \%>\%
  str()
}
