% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{productSize}
\alias{productSize}
\title{Set product size}
\usage{
productSize(settings, range)
}
\arguments{
\item{settings}{a settings list created by `p3Settings`}

\item{range}{a vector with two numbers, c(minimal, maximal) product size
(inclusive).}
}
\value{
a settings list with updated values
}
\description{
Set product size
}
\examples{
str(p3Settings() \%>\% productSize(c(100,300)))

}
