% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selectPairs}
\alias{selectPairs}
\title{Given a list of exons returns candidates for primer design}
\usage{
selectPairs(exonsByTx, tolerance = 1)
}
\arguments{
\item{exonsByTx}{exon GRanges by transcripts}

\item{tolerance}{the minimal tolerable difference in splice junctions 
coordinates (0: return everything, 1: only unique splice junctions).}
}
\value{
a GRangesList
}
\description{
The transcripts of single exon are excluded!
}
