% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{toGRanges}
\alias{toGRanges}
\title{Crate GRanges object from the Primer3 output}
\usage{
toGRanges(primers, exons)
}
\arguments{
\item{primers}{}
}
\description{
Crate GRanges object from the Primer3 output
}
