% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{diagnose}
\alias{diagnose}
\title{Returns diagnostics information from the run}
\usage{
diagnose(result)
}
\arguments{
\item{result}{a value returned by \link{runPrimer3}}
}
\value{
a list with statistics of the run for the left, right primers
 and their pairs.
}
\description{
For this to work, `PRIMER_EXPLAIN_FLAG` in the must be set to "1".
}
\examples{
# Too short sequence
seq1 <- paste(c("A", "T", "G", "C")[sample(4, 100, replace = TRUE)], 
              collapse = "")
seqOpts <- seqSettings(seqId = "seq1", seq = seq1)
res <- design(seqOpts)
diagnose(res)

}
