% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{toGRanges}
\alias{toGRanges}
\title{Transforms output of design to GRanges}
\usage{
toGRanges(primers, exons)
}
\arguments{
\item{primers}{the "primers" item of the `\link{design}` output}

\item{exons}{the GRanges of the exons}
}
\value{
GRangesList
}
\description{
Transforms output of design to GRanges
}
